A little fix...

+++ RX.h
@@ -325,7 +325,7 @@ void loop()
 
   // sample RSSI when packet is in the 'air'
   if ((lostpack < 2) && (last_rssi_time != last_pack_time) &&
-      (time - last_pack_time) > (uint32_t)(modem_params[bind_data.modem_params].interval - 1500)) {
+      (time - last_pack_time) > (modem_params[bind_data.modem_params].interval - 1500)) {
     last_rssi_time = last_pack_time;
     last_rssi_value = rfmGetRSSI(); // Read the RSSI value
     RSSI_sum += last_rssi_value;    // tally up for average
@@ -342,12 +342,12 @@ void loop()
   time = micros();
 
   if (firstpack) {
-    if ((!lostpack) && (time - last_pack_time) > (uint32_t)(modem_params[bind_data.modem_params].interval + 1000)) {
+    if ((!lostpack) && (time - last_pack_time) > (modem_params[bind_data.modem_params].interval + 1000)) {
       // we missed one packet, hop to next channel
       lostpack = 1;
       last_pack_time += modem_params[bind_data.modem_params].interval;
       willhop = 1;
-    } else if ((lostpack == 1) && (time - last_pack_time) > (uint32_t)(modem_params[bind_data.modem_params].interval + 1000)) {
+    } else if ((lostpack == 1) && (time - last_pack_time) > (modem_params[bind_data.modem_params].interval + 1000)) {
       // we lost second packet in row, hop and signal trouble
       lostpack = 2;
       last_pack_time += modem_params[bind_data.modem_params].interval;
